fix: require X-API-Key auth on all ML routers and stop host-exposing the ML engine port
