+++ backend/ml/main.py
@@ -1,11 +1,10 @@
 import asyncio
-import hmac
 import logging
 import os
 import time
 import numpy as np
 from datetime import datetime, timedelta
-from fastapi import FastAPI, HTTPException, Header, Depends, Query
+from fastapi import FastAPI, HTTPException, Depends, Query
 from fastapi.middleware.cors import CORSMiddleware
 from pydantic import BaseModel, Field
 from typing import List, Optional
@@ -27,7 +26,7 @@ from app.models.mid_trip_reoptimiser import find_mid_trip_loads
 from app.models.base import model_exists
 from app.models.demand_forecast import MODEL_NAME as DEMAND_MODEL_NAME
 from app.models.price_prediction import MODEL_NAME as PRICE_MODEL_NAME
-from routes import register_ml_routers
+from routes import register_ml_routers, verify_api_key
 
 logging.basicConfig(
     level=logging.INFO,
@@ -38,21 +37,13 @@ logger = logging.getLogger(__name__)
 # Track loaded models for health reporting
 loaded_models: set[str] = set()
 
-async def verify_api_key(x_api_key: str = Header(None, alias="X-API-Key")):
-    ml_api_key = os.environ.get("ML_API_KEY")
-    if not ml_api_key:
-        logger.warning("ML_API_KEY not set - ML engine is unavailable (503)")
-        raise HTTPException(status_code=503, detail="ML engine not configured: missing ML_API_KEY")
-    if not x_api_key or not hmac.compare_digest(x_api_key, ml_api_key):
-        raise HTTPException(status_code=401, detail="Unauthorized")
-
-
 app = FastAPI(
     title="Truxify ML Engine",
     description="ML prediction service for load matching, pricing, ETA, and route optimization",
     version="1.0.0",
-    docs_url="/docs",      # Swagger UI at /docs
-    redoc_url="/redoc", 
+    # Swagger/ReDoc interactive docs are disabled in production.
+    docs_url=None if os.environ.get("ENVIRONMENT") == "production" else "/docs",
+    redoc_url=None if os.environ.get("ENVIRONMENT") == "production" else "/redoc",
 )
